+++ cryptofeed/exchanges/kraken_futures.py
@@ -1,5 +1,5 @@
 '''
-Copyright (C) 2017-2022 Bryant Moscon - bmoscon@gmail.com
+Copyright (C) 2017-2023 Bryant Moscon - bmoscon@gmail.com
 
 Please see the LICENSE file for the terms and conditions
 associated with this software.
@@ -63,7 +63,7 @@ class KrakenFutures(Feed):
                 stype = FUTURES
                 symbol, expiry = symbol.split("_")
             symbol = symbol.replace('XBT', 'BTC')
-            base, quote = symbol[:3], symbol[3:]
+            base, quote = symbol[:len(symbol) - 3], symbol[-3:]
 
             s = Symbol(base, quote, type=stype, expiry_date=expiry)