+++ tests/unit/test_binance.py
@@ -1,14 +1,17 @@
 '''
-Copyright (C) 2017-2022 Bryant Moscon - bmoscon@gmail.com
+Copyright (C) 2017-2023 Bryant Moscon - bmoscon@gmail.com
 
 Please see the LICENSE file for the terms and conditions
 associated with this software.
 '''
 import random
 
+import pytest
+
 from cryptofeed.exchanges import Binance
 
 
+@pytest.mark.xfail(reason="Binance blocks build machine IP ranges. If not in the USA this should pass")
 def test_binance_address_generation():
     symbols = Binance.symbols()
     channels = [channel for channel in Binance.info()['channels']['websocket'] if not Binance.is_authenticated_channel(channel)]