+++ tools/generate_test_data.py
@@ -1,5 +1,5 @@
 '''
-Copyright (C) 2017-2022 Bryant Moscon - bmoscon@gmail.com
+Copyright (C) 2017-2023 Bryant Moscon - bmoscon@gmail.com
 
 Please see the LICENSE file for the terms and conditions
 associated with this software.
@@ -13,7 +13,7 @@ import uvloop
 from cryptofeed.feedhandler import FeedHandler
 from cryptofeed.exchanges import EXCHANGE_MAP
 from cryptofeed.raw_data_collection import AsyncFileCallback
-from cryptofeed.defines import BINANCE, BINANCE_FUTURES, BINANCE_US, BITFINEX, L2_BOOK, TRADES, TICKER, CANDLES, EXX
+from cryptofeed.defines import BINANCE, BINANCE_FUTURES, BINANCE_US, BINANCE_TR, BITFINEX, L2_BOOK, TRADES, TICKER, CANDLES, EXX
 from check_raw_dump import main as check_dump
 
 asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
@@ -43,7 +43,7 @@ def main(only_exchange=None):
         info = exchange.info()
         channels = list(set.intersection(set(info['channels']['websocket']), set([L2_BOOK, TRADES, TICKER, CANDLES])))
         sample_size = 10
-        if exch_str in (BINANCE_US, BINANCE):
+        if exch_str in (BINANCE_US, BINANCE_TR, BINANCE):
             # books of size 5000 count significantly against rate limits
             sample_size = 4
         while True:
@@ -82,4 +82,4 @@ def main(only_exchange=None):
 
 
 if __name__ == '__main__':
-    main()
+    main('OKCOIN')