+++ wheels.sh
@@ -1,0 +1,1 @@
+docker run --rm -v `pwd`:/io quay.io/pypa/manylinux_2_24_x86_64 /io/build-wheels.sh