+++ CHANGES.md
@@ -1,11 +1,24 @@
 ## Changelog
 
-### 2.3.2
+### 2.4.0
+ * Update: Fix tests
+ * Update: Okcoin moved to v5 API used by OKX
+ * Bugfix: InfluxDB none type conversions
+ * New Exchange: GateIO Futures
+ * Feature: Added support for Bybit spot orderbook and trade websocket endpoints
+
+### 2.3.2 (2023-05-27)
+ * Bugfix: Fix Socket backend
  * Bugfix: Fix AUCTION symbol parsing on Coinbase
  * Bugfix: Fix PERPETUAL symbol parsing on Phemex
- * Feature: Added support for Bybit spot orderbook and trade websocket endpoints
+ * Bugfix: Fix PERPETUAL symbol parsing on Kraken Futures
  * Feature: Access to all AIOKafka configuration options
  * Feature: Use backend Queue for Kafka
+ * Feature: Add support for storing book snapshots in Redis as key-value
+ * Update: Switch from unmaintained aioredis to redis-py
+ * Bugfix: Correct value for Crypto.com Ask price
+ * Update: Remove cChardet dependency
+ * Feature: Binance TR support
 
 ### 2.3.1 (2022-10-31)
  * Bugfix: timestamp not reset correctly on reconnect